backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.57
python-dotenv==1.0.0
watchtower==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
